Move shared code to 'TestUtils'
